Make test_decorator.py compliant with pylint 1.7
The ignored-argument-names option is now used for ignoring arguments
for unused-variable check, so the name "dummy" (previously ignored
because of the dummy-variables-rgx option) results in a warning.

This in turn leads to a score less than 10 and fails the tox run.

Name with a leading underscore can be used to suppress the warning.