Omit the badges line from the long description
PyPI cannot use relative paths to changelog or license file, nor can it
refer to the develop branch build status.

This makes the badges rather useless here.

As a future enhancement, I might consider keeping README,rst in the git
repository instead of building it from README.md.  This way I will be
able to refer to the relevant (master) branch and specific revisions of
the files at github.